Restructure
:P
